requirements.txt
paho-mqtt==1.5.1
pyserial==3.5
python-dotenv==0.17.0
